core/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
underthesea>=1.3.1
flask==1.1.2
gunicorn==20.0.4
torch #==1.6.0
torchvision #==0.7.0
joblib #==1.0.0
requests==2.25.1
Pillow>=8.1.0
opencv-python-headless
scikit-learn==0.24.1
pytorch-pretrained-bert
timeout-decorator
beautifulsoup4
google-api-python-client==1.7.11
nltk
uuid
tqdm
lxml
tensorflow #<2.4